refactor(maps): extract the shared map-texture candidate resolver
HuntZoneMapTextures and HousingGameMaps each carried their own copy of
TextureCandidates/FileExists/ResolveTexturePath, with Hunts silently
missing the _s.tex fallback candidate Housing had. Extract both into
Core/Maps/MapTextures, keeping all five candidates so the two callers
stay behaviorally identical; both now pass a subsystem label through
for the existing Debug-level 'could not test' log line.
